Remove the dependency that DataOffset and ContractionTableOffset need to be the same. (dm)
git-svn-id: svn://mielke.cc/main/brltty@3775 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/ctb_internal.h
@@ -27,7 +27,7 @@ extern "C" {
 #define HASHNUM 1087
 #define CTH(x) (((x[0]<<8)+x[1])%HASHNUM)
 
-typedef unsigned long int ContractionTableOffset;
+typedef uint32_t ContractionTableOffset;
 
 typedef enum {
   CTC_Space       = 0X01,
